requirements.txt
requests~=2.31.0
python-gitlab~=4.4.0
dnspython~=2.4.2